Add a SceneformManager and remove CameraXAPI logic in MainActivity.
